fix: CPU build fixes for template params, JavaCPP parsing, and Maven compiler
- Fix sd_tanh/sd_cos/sd_sin template calls to use two params <X, Z>
- Remove invalid BUILD_SINGLE_TEMPLATE on static functions (6 files)
- Fix OMP reduction syntax in distillation_kl_loss.cpp
- Fix OpDescriptor.h: hide OpTraits enum from JavaCPP, remove SD_PADDED_NEW_DELETE
- Fix MemoryConfig.h: remove __JAVACPP_HACK__ guards matching other config headers
- Add preset Info entries for SD_VALIDATE_PTR/OpTraits in CPU and CUDA presets
- Fix nd4j-native pom.xml: disable --release to resolve --add-exports conflict

+++ libnd4j/include/graph/impl/Node.cpp
@@ -531,8 +531,7 @@ sd::graph::Node::Node(const ::graph::FlatNode* node) {
       if (node->extraInteger() != nullptr)
         for (uint32_t e = 0; e < node->extraInteger()->size(); e++) {
           auto v = node->extraInteger()->Get(e);
-          // FIXME: remove this static_cast, iArgs should be sd::LongType
-          block->getIArguments()->emplace_back(static_cast<int>(v));
+          block->getIArguments()->emplace_back(v);
         }
 
       if (node->extraParams() != nullptr)
